Put all wire-format related helpers on ProtoWriter.
